backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.8.0,<3.0.0
requests>=2.31.0,<3.0.0
neo4j>=5.21.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0

